# app/swagger/swagger.json
{
  "components": {
...
    "schemas": {
...
      "UpdateMemberStatusDetails": {
        "properties": {
          "membContribDetlId": {
            "type": "string"
          },
          "schdlMembStatusCd": {
            "type": "string"
          }
        },
        "required": ["membContribDetlId", "schdlMembStatusCd"],
        "type": "object",
        "additionalProperties": false
      },
...
      "FilterOptions": {
        "properties": {
          "limit": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "ContributionHeaderUpdateError": {
        "properties": {
          "errorCode": {
            "type": "string"
          },
          "errorDetail": {
            "type": "string"
          }
        },
        "required": ["errorCode", "errorDetail"],
        "type": "object",
        "additionalProperties": false
      },
...
      "MemberContributionDetailsResponse": {
        "properties": {
          "membContribDetlId": {
            "type": "string"
          },
          "nestScheduleRef": {
            "type": "string"
          },
          "membEnrolmentRef": {
            "type": "string"
          },
          "membContriDueDate": {
            "type": "string",
            "format": "date"
          },
          "membPlanRef": {
            "type": "string"
          },
          "empGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "groupName": {
            "type": "string"
          },
          "schdlMembStatusCd": {
            "type": "string"
          },
          "membPartyId": {
            "type": "string"
          },
          "scmPartyId": {
            "type": "string"
          },
          "nino": {
            "type": "string"
          },
          "alternativeId": {
            "type": "string"
          },
          "autoCalcFlag": {
            "type": "string"
          },
          "pensEarnings": {
            "type": "number",
            "format": "float"
          },
          "emplContriAmt": {
            "type": "number",
            "format": "float"
          },
          "membContriAmt": {
            "type": "number",
            "format": "float"
          },
          "membNonPayReason": {
            "type": "string"
          },
          "membLeaveEarnings": {
            "type": "number",
            "format": "float"
          },
          "newGroupName": {
            "type": "string"
          },
          "newGroupPensEarnings": {
            "type": "number",
            "format": "float"
          },
          "newGroupEmplContriAmt": {
            "type": "number",
            "format": "float"
          },
          "newGroupMembContriAmt": {
            "type": "number",
            "format": "float"
          },
          "optoutRefNum": {
            "type": "string"
          },
          "optoutDeclarationFlag": {
            "type": "string"
          },
          "newPaymentPlanNo": {
            "type": "string"
          },
          "newPaymentSourceName": {
            "type": "string"
          },
          "membNonPayEffDate": {
            "type": "string",
            "format": "date"
          },
          "secEnrolPensEarnings": {
            "type": "number",
            "format": "float"
          },
          "secEnrolEmplContriAmt": {
            "type": "number",
            "format": "float"
          },
          "secEnrolMembContriAmt": {
            "type": "number",
            "format": "float"
          },
          "channelType": {
            "type": "string"
          },
          "memberExcludedFlag": {
            "type": "string"
          },
          "membPaymentDueDate": {
            "type": "string",
            "format": "date"
          },
          "newEmpGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "recordStartDate": {
            "type": "string",
            "format": "date"
          },
          "recordEndDate": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "employerNestId": {
            "type": "string"
          },
          "origScheduleRef": {
            "type": "string"
          },
          "contribHeaderId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "enrolmentType": {
            "type": "string"
          },
          "secEnrolmentType": {
            "type": "string"
          },
          "emplContriPct": {
            "type": "number",
            "format": "float"
          },
          "membContriPct": {
            "type": "number",
            "format": "float"
          },
          "newGroupEmplContriPct": {
            "type": "number",
            "format": "float"
          },
          "newGroupMembContriPct": {
            "type": "number",
            "format": "float"
          },
          "recordChangedFlag": {
            "type": "string"
          },
          "memTaxReliefEligibility": {
            "type": "string"
          },
          "origMembNonPayReason": {
            "type": "string"
          }
        },
        "required": [
          "membContribDetlId",
          "nestScheduleRef",
          "membEnrolmentRef",
          "membContriDueDate",
          "empGroupId",
          "groupName",
          "schdlMembStatusCd",
          "membPartyId",
          "membNonPayReason",
          "firstName",
          "lastName",
          "emplContriPct",
          "membContriPct"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "RetriveContributionDetailsResponse": {
        "properties": {
          "contribHeaderId": {
            "type": "string"
          },
          "origScheduleRef": {
            "type": "string"
          },
          "employerNestId": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          },
          "membPaymentDueDate": {
            "type": "string",
            "format": "date"
          },
          "recordEndDate": {
            "type": "string",
            "format": "date"
          },
          "createdBy": {
            "type": "string"
          },
          "newEmpGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "recordStartDate": {
            "type": "string",
            "format": "date"
          },
          "memberExcludedFlag": {
            "type": "string"
          },
          "nestScheduleRef": {
            "type": "string"
          },
          "membEnrolmentRef": {
            "type": "string"
          },
          "membContribDetlId": {
            "type": "string"
          },
          "membPlanRef": {
            "type": "string"
          },
          "membContriDueDate": {
            "type": "string",
            "format": "date"
          },
          "empGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "groupName": {
            "type": "string"
          },
          "schdlMembStatusCd": {
            "type": "string"
          },
          "membPartyId": {
            "type": "string"
          },
          "scmPartyId": {
            "type": "string"
          },
          "nino": {
            "type": "string"
          },
          "alternativeId": {
            "type": "string"
          },
          "autoCalcFlag": {
            "type": "string"
          },
          "emplContriAmt": {
            "type": "number",
            "format": "float"
          },
          "pensEarnings": {
            "type": "number",
            "format": "float"
          },
          "membContriAmt": {
            "type": "number",
            "format": "float"
          },
          "membLeaveEarnings": {
            "type": "number",
            "format": "float"
          },
          "membNonPayReason": {
            "type": "string"
          },
          "newGroupPensEarnings": {
            "type": "number",
            "format": "float"
          },
          "newGroupName": {
            "type": "string"
          },
          "optoutDeclarationFlag": {
            "type": "string"
          },
          "newGroupEmplContriAmt": {
            "type": "number",
            "format": "double"
          },
          "optoutRefNum": {
            "type": "string"
          },
          "newGroupMembContriAmt": {
            "type": "number",
            "format": "float"
          },
          "newPaymentPlanNo": {
            "type": "string"
          },
          "secEnrolMembContriAmt": {
            "type": "number",
            "format": "float"
          },
          "newPaymentSourceName": {
            "type": "string"
          },
          "membNonPayEffDate": {
            "type": "string",
            "format": "date"
          },
          "secEnrolPensEarnings": {
            "type": "number",
            "format": "float"
          },
          "channelType": {
            "type": "string"
          },
          "enrolmentType": {
            "type": "string"
          },
          "emplContriPct": {
            "type": "number",
            "format": "float"
          },
          "membContriPct": {
            "type": "number",
            "format": "float"
          },
          "newGroupEmplContriPct": {
            "type": "number",
            "format": "float"
          },
          "secEnrolEmplContriAmt": {
            "type": "number",
            "format": "double"
          },
          "firstName": {
            "type": "string"
          },
          "secEnrolmentType": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "newGroupMembContriPct": {
            "type": "number",
            "format": "float"
          },
          "recordChangedFlag": {
            "type": "string"
          },
          "origMembNonPayReason": {
            "type": "string"
          },
          "memTaxReliefEligibility": {
            "type": "string"
          },
          "errorDetails": {
            "$ref": "#/components/schemas/ErrorDetailsResponse"
          },
          "rdschedulememberstatus": {
            "$ref": "#/components/schemas/RDScheduleMemberStatusResponse"
          },
          "rdpartcontribreason": {
            "$ref": "#/components/schemas/RDPartContribReasonResponse"
          }
        },
        "required": [
          "nestScheduleRef",
          "membEnrolmentRef",
          "membContribDetlId",
          "membContriDueDate",
          "empGroupId",
          "groupName",
          "schdlMembStatusCd",
          "membPartyId",
          "membNonPayReason",
          "emplContriPct",
          "membContriPct",
          "firstName",
          "lastName",
          "errorDetails",
          "rdschedulememberstatus",
          "rdpartcontribreason"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "SearchMemberContributionResultResponse_RetriveContributionDetailsResponse_": {
        "properties": {
          "totalRecordCount": {
            "type": "number",
            "format": "double"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RetriveContributionDetailsResponse"
            },
            "type": "array"
          }
        },
        "required": ["totalRecordCount", "results"],
        "type": "object",
        "additionalProperties": false
      },
...
      "DetailsFilterParams": {
        "properties": {
          "nestScheduleRef": {
            "type": "string"
          },
          "employerNestId": {
            "type": "string"
          },
          "contribHeaderId": {
            "type": "string"
          },
          "fitstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": ["nestScheduleRef", "fitstName", "lastName"],
        "type": "object",
        "additionalProperties": false
      },
      "DetailsFilterElements": {
        "properties": {
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          },
          "params": {
            "$ref": "#/components/schemas/DetailsFilterParams"
          }
        },
        "required": ["params"],
        "type": "object",
        "additionalProperties": false
      },
...
      "ContributionDetailsResponseDTO": {
        "properties": {
          "message": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ContributionHeaderUpdateError"
            },
            "type": "array"
          },
          "type": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
...
      "UpdateMemberStatusRequest": {
        "properties": {
          "contributionDetail": {
            "items": {
              "$ref": "#/components/schemas/UpdateMemberStatusDetails"
            },
            "type": "array"
          }
        },
        "required": ["contributionDetail"],
        "type": "object",
        "additionalProperties": false
      },
      "ContributionMemberDetails": {
        "description": "This is Contribution member details interface",
        "properties": {
          "membContribDetlId": {
            "type": "string"
          },
          "schdlMembStatusCd": {
            "type": "string"
          },
          "membPartyId": {
            "type": "string"
          },
          "scmPartyId": {
            "type": "string"
          },
          "nino": {
            "type": "string"
          },
          "alternativeId": {
            "type": "string"
          },
          "autoCalcFlag": {
            "type": "string"
          },
          "pensEarnings": {
            "type": "number",
            "format": "double"
          },
          "emplContriAmt": {
            "type": "number",
            "format": "double"
          },
          "membContriAmt": {
            "type": "number",
            "format": "double"
          },
          "membNonPayReason": {
            "type": "string"
          },
          "membLeaveEarnings": {
            "type": "number",
            "format": "double"
          },
          "newEmpGroupId": {
            "type": "number",
            "format": "double"
          },
          "newGroupName": {
            "type": "string"
          },
          "newGroupPensEarnings": {
            "type": "number",
            "format": "double"
          },
          "newGroupEmplContriAmt": {
            "type": "number",
            "format": "double"
          },
          "newGroupMembContriAmt": {
            "type": "number",
            "format": "double"
          },
          "optoutRefNum": {
            "type": "string"
          },
          "optoutDeclarationFlag": {
            "type": "string"
          },
          "newPaymentPlanNo": {
            "type": "string"
          },
          "newPaymentSourceName": {
            "type": "string"
          },
          "membNonPayEffDate": {
            "type": "string",
            "format": "date"
          },
          "secEnrolPensEarnings": {
            "type": "number",
            "format": "double"
          },
          "secEnrolEmplContriAmt": {
            "type": "number",
            "format": "double"
          },
          "secEnrolMembContriAmt": {
            "type": "number",
            "format": "double"
          },
          "channelType": {
            "type": "string"
          },
          "memberExcludedFlag": {
            "type": "string"
          },
          "membPaymentDueDate": {
            "type": "string",
            "format": "date"
          },
          "recordStartDate": {
            "type": "string",
            "format": "date"
          },
          "recordEndDate": {
            "type": "string",
            "format": "date"
          },
          "emplContriPct": {
            "type": "number",
            "format": "double"
          },
          "membContriPct": {
            "type": "number",
            "format": "double"
          },
          "newGroupEmplContriPct": {
            "type": "number",
            "format": "double"
          },
          "recordChangedFlag": {
            "type": "string"
          },
          "memTaxReliefEligibility": {
            "type": "string"
          },
          "origMembNonPayReason": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        },
        "required": ["membContribDetlId"],
        "type": "object",
        "additionalProperties": false
      },
      "ContributionDetailsRequestDTO": {
        "description": "This is Contribution member details update API Request DTO interface",
        "properties": {
          "contributionDetail": {
            "items": {
              "$ref": "#/components/schemas/ContributionMemberDetails"
            },
            "type": "array"
          }
        },
        "required": ["contributionDetail"],
        "type": "object",
        "additionalProperties": false
      },
...
      "SearchResultsetResponse_MemberContributionDetailsResponse_": {
        "properties": {
          "totalRecordCount": {
            "type": "number",
            "format": "double"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/MemberContributionDetailsResponse"
            },
            "type": "array"
          }
        },
        "required": ["totalRecordCount", "results"],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "x-functions-key",
        "in": "header"
      }
    }
  },
  "info": {
    "title": "Contribution Index APIs",
    "version": "1.63.0",
    "description": "REST APIs for Contribution Index",
    "contact": {}
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/contribution/retrieveContributionDetails": {
      "post": {
        "operationId": "GetDetailsByFilter",
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SearchMemberContributionResultResponse_RetriveContributionDetailsResponse_"
                    },
                    {}
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request"
          },
          "404": {
            "description": "Record does not exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "description": "5205 API Catalogue Number\r\nRetrieves a list of details with filter criterias",
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetailsFilterElements"
              }
            }
          }
        }
      }
    },
    "/contribution/contributionDetails": {
      "put": {
        "operationId": "UpdateMemberContributionDetails",
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ContributionDetailsResponseDTO"
                    },
                    {}
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request"
          },
          "404": {
            "description": "Record does not exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "description": "5204 API Catalogue Number\r\nUpdate the Contribution Details",
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionDetailsRequestDTO"
              }
            }
          }
        }
      }
    },
    "/contribution/Submission/MemberContributionSubmissions/{contribSubmissionRef}": {
      "get": {
        "operationId": "GetMemberContributionSubmission",
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SearchResultsetResponse_MemberContributionDetailsResponse_"
                    },
                    {}
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid Request"
          },
          "404": {
            "description": "Record does not exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "description": "5404 API Catalogue Number\r\nRetrieves returns a list of all Member Contribution Submissions based on contribSubmissionRef passed.",
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [
          {
            "description": "contribSubmissionRef of the Member Contribution Submission record to be fetched",
            "in": "path",
            "name": "contribSubmissionRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/contribution/updateMemberStatus": {
      "put": {
        "operationId": "UpdateMemberStatus",
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid Request"
          },
          "404": {
            "description": "Record does not exist"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "description": "5207 API Catalogue Number Update member status",
        "security": [
          {
            "api_key": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "description": "Array of the Member Contribution Details Id & Member Status Array<Contribution_Details>",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberStatusRequest"
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:7071/api"
    }
  ]
}